+++ src/BasicBehaveEngine/nodes/customEvent/Receive.ts
@@ -1,5 +1,5 @@
 import {BehaveEngineNode, IBehaviourNodeProps} from "../../BehaveEngineNode";
-import {IInteractivityEvent} from "../../../types/InteractivityGraph";
+import {IInteractivityEvent} from "../../types/InteractivityGraph";
 
 export class Receive extends BehaveEngineNode {
     REQUIRED_CONFIGURATIONS = {event: {}}
