+++ src/BasicBehaveEngine/nodes/math/arithmetic/AbsoluteValue.ts
@@ -1,4 +1,4 @@
-import { IInteractivityValue } from "../../../../types/InteractivityGraph";
+import { IInteractivityValue } from "../../../types/InteractivityGraph";
 import {BehaveEngineNode, IBehaviourNodeProps} from "../../../BehaveEngineNode";
 
 export class AbsoluteValue extends BehaveEngineNode {
