+++ src/BasicBehaveEngine/nodes/pointer/PointerGet.ts
@@ -1,4 +1,4 @@
-import { IInteractivityValue } from "../../../types/InteractivityGraph";
+import { IInteractivityValue } from "../../types/InteractivityGraph";
 import {BehaveEngineNode, IBehaviourNodeProps} from "../../BehaveEngineNode";
 
 export class PointerGet extends BehaveEngineNode {
